+++ containers/Login/LoginContainer.tsx
@@ -1,31 +1,15 @@
-import { useRouter } from 'next/router';
-import { useEffect } from 'react';
+import { styled } from 'styled-components';
+
+import { Login } from '@/components/Login';
 
 const LoginContainer = () => {
-  const router = useRouter();
-
-  useEffect(() => {
-    const params = new URL(document.location.toString()).searchParams;
-    const code = params.get('code');
-    const token = params.get('token');
-
-    /** in case of user has problem with kakao auth server */
-    if (token === null) {
-      console.log(params);
-      console.log(code);
-      router.push('/');
-      return;
-    }
-
-    try {
-      console.log(token);
-      router.push('/home');
-    } catch (err) {
-      console.error(err);
-      router.push('/');
-    }
-  }, []);
-  return <p>로그인 페이지</p>;
+  return (
+    <StLoginWrapper>
+      <Login />
+    </StLoginWrapper>
+  );
 };
 
 export default LoginContainer;
+
+const StLoginWrapper = styled.main``;
