+++ package.json
@@ -15,6 +15,7 @@
     "eslint-config-next": "13.4.13",
     "firebase": "^10.4.0",
     "firebase-tools": "^12.5.4",
+    "jwt-decode": "^3.1.2",
     "mediasoup-client": "3.6.71",
     "next": "13.4.13",
     "next-pwa": "^5.6.0",
@@ -46,7 +47,7 @@
   "private": true,
   "scripts": {
     "build": "next build",
-    "dev": "next dev -p 3013",
+    "dev": "next dev -p 3004",
     "lint": "next lint",
     "start": "next start"
   },
